Fix producer cache key for JSON schema (#137)
The key used to cache producers includes the producer schema.
This works well for primitive schema types and for complex
schema types when the schema is the exact same instance. However, the
schema contract does not enforce equals/hashCode. This results in cases
where 2 producers using the same type of schema are not matching in the
cache (eg. Schema.JSON(<some-class>)). This commit fixes this by using
the schema hash for matching rather than the schema itself.