Update Pulsar to 3.3.2 (#189)
* Update Pulsar to 3.3.2

* Fix compilation in test due to Pulsar 3.3.2 update

* Replace order of `given/willAnswer` for spied objects

The cause of the test hang was that the test was incorrectly setting up
the spy on the type message builder impl. In previous Pulsar version of
TypedMessageBuilderImpl, the fact that the method sendAsync was being
called at mock setup time was not causing an issue. However, in the
latest impl it did not like that and was throwing things off.

Spied objects should always use the `doReturn|Answer|Throw()` family
as described in https://javadoc.io/doc/org.mockito/mockito-core/latest/org/mockito/Mockito.html#important-gotcha-on-spying-real-objects--heading

---------

Co-authored-by: onobc <onobc@apache.org>